cleanup polars-time rolling window dispatch
