+++ crates/polars-time/src/prelude.rs
@@ -1,7 +1,7 @@
 pub use date_range::*;
 
 pub use crate::chunkedarray::*;
-pub use crate::series::{SeriesOpsTime, TemporalMethods};
+pub use crate::series::TemporalMethods;
 pub use crate::windows::bounds::*;
 pub use crate::windows::duration::*;
 pub use crate::windows::group_by::*;
