--- crates/polars-time/src/series/implementations/date.rs
@@ -1,7 +1,0 @@
-use super::*;
-
-impl SeriesOpsTime for Wrap<DateChunked> {
-    fn ops_time_dtype(&self) -> &DataType {
-        self.0.dtype()
-    }
-}
