--- crates/polars-time/src/series/implementations/list.rs
@@ -1,7 +1,0 @@
-use super::*;
-
-impl SeriesOpsTime for Wrap<ListChunked> {
-    fn ops_time_dtype(&self) -> &DataType {
-        self.0.dtype()
-    }
-}
