--- crates/polars-time/src/series/implementations/struct_.rs
@@ -1,7 +1,0 @@
-use super::*;
-
-impl SeriesOpsTime for Wrap<StructChunked> {
-    fn ops_time_dtype(&self) -> &DataType {
-        self.0.dtype()
-    }
-}
