--- crates/polars-time/src/series/implementations/time.rs
@@ -1,7 +1,0 @@
-use super::*;
-
-impl SeriesOpsTime for Wrap<TimeChunked> {
-    fn ops_time_dtype(&self) -> &DataType {
-        self.0.dtype()
-    }
-}
