--- crates/polars-time/src/series/implementations/utf8.rs
@@ -1,7 +1,0 @@
-use super::*;
-
-impl SeriesOpsTime for Wrap<Utf8Chunked> {
-    fn ops_time_dtype(&self) -> &DataType {
-        self.0.dtype()
-    }
-}
